% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_cache}
\alias{address_cache}
\title{address_cache}
\usage{
address_cache(
  address_stream,
  cache_name,
  .function,
  key_cols = as.character(address_cols),
  db_name = tessilake::cache_path("address_stream.sqlite", "deep", "stream"),
  ...
)
}
\arguments{
\item{address_stream}{data.table of addresses}

\item{cache_name}{name of cache file}

\item{.function}{function to be called for processing, is sent \code{address_stream[address_cols]} and additional parameters.}

\item{key_cols}{as.character(address_cols)}

\item{db_name}{path to sqlite database, defaults to \code{tessilake::cache_path("address_stream.sqlite","deep","stream")}}

\item{...}{additional options passed to \code{.function}}
}
\value{
data.table of addresses processed
}
\description{
Handles caching of already-processed addresses so that they're only processed once
}
