% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_census.R
\name{address_census}
\alias{address_census}
\title{address_census}
\usage{
address_census(address_stream)
}
\arguments{
\item{address_stream}{data.table of addresses}
}
\value{
data.table of addresses, one row per input row. Contains \code{address_cols}, \code{timestamp} and added census features.
}
\description{
Append census features to a data.table of addresses based on matching \code{address_cols} and \code{timestamp}
}
