% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_clean}
\alias{address_clean}
\title{address_clean}
\usage{
address_clean(address_stream)
}
\arguments{
\item{address_stream}{data.table of addresses}
}
\value{
data.table of addresses cleaned
}
\description{
Removes newlines, tabs, lowercases, trims whitespace, and removes junk info
}
