% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_create_stream}
\alias{address_create_stream}
\title{address_create_stream}
\usage{
address_create_stream(...)
}
\arguments{
\item{...}{additional parameters passed on to \link{stream_from_audit}}
}
\value{
data.table of addresses data at different points of time, no more than one
change per address per day
}
\description{
Creates address data with timestamps from \code{TA_AUDIT_TABLE} and \code{T_ADDRESS} data
}
