% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_geocode.R
\name{address_exec_census}
\alias{address_exec_census}
\title{address_exec_census}
\usage{
address_exec_census(address_stream)
}
\arguments{
\item{address_stream}{data.table of addressses, containing at least \code{census.street}, \code{census.city}, \code{census.state}, \code{census.postal_code}, \code{census.vintageName}}
}
\value{
data.table with census geocoding data appended as additional columns
}
\description{
Wrapper for \link[censusxy:cxy_geocode]{censusxy::cxy_geocode}
}
