% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_exec_libpostal}
\alias{address_exec_libpostal}
\title{address_exec_libpostal}
\usage{
address_exec_libpostal(addresses)
}
\arguments{
\item{addresses}{character vector of addresses}
}
\value{
data.frame of parsed addresses
}
\description{
address_exec_libpostal
}
