% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_geocode.R
\name{address_geocode_all}
\alias{address_geocode_all}
\alias{address_geocode}
\title{address_geocode}
\usage{
address_geocode_all(address_stream, num_tries = 10)

address_geocode(address_stream)
}
\arguments{
\item{address_stream}{data.table of addresses}

\item{num_tries}{integer number of times to retry on error}
}
\value{
data.table of addresses, one row per address in input, must include \code{address_cols}. Contains only address_cols and columns returned by \code{tidygeocoder}
}
\description{
Geocodes addresses using the \link{tidygeocoder} package.
Tries each address up to six times, using \code{libpostal} parsing (if \code{libpostal.house_number} and \code{libpostal.road} are passed in as columnes),
\code{street1}, and \code{street2}, and the US census and openstreetmap geocoders.
}
\section{Functions}{
\itemize{
\item \code{address_geocode_all()}: geocode all addresses

\item \code{address_geocode()}: geocode only uncached addresses, load others from cache

}}
