% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_geocode.R
\name{address_geocode_all}
\alias{address_geocode_all}
\title{address_geocode_all}
\usage{
address_geocode_all(address_stream)
}
\arguments{
\item{address_stream}{data.table of addresses, must include \code{address_cols}}
}
\value{
data.table of addresses, one row address in \code{address_stream}
}
\description{
Geocodes addresses using the \link{tidygeocoder} package . Tries each address up to six times, using \code{libpostal} parsing, \code{street1}, and \code{street2}, and the
US census and openstreetmap geocoders.
}
