% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_geocode.R
\name{address_geocode_census}
\alias{address_geocode_census}
\title{address_geocode_census}
\usage{
address_geocode_census(address_stream)
}
\arguments{
\item{address_stream}{data.table of addresses, must include \code{address_cols} and \code{timestamp}}
}
\value{
data.table of addresses, one row per unique geocoded address, with census geocode information appended
}
\description{
Geocodes addresses using the US census geocoding API. Tries each address up to three times, using \code{libpostal} parsing, \code{street1}, and \code{street2}.
}
