% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/street_cleaner.R
\name{address_normalize}
\alias{address_normalize}
\title{address_normalize}
\usage{
address_normalize(address_stream)
}
\arguments{
\item{address_stream}{data.table of addresses}
}
\value{
data.table of addresses parsed, one per row in address_stream. Contains only address_cols and address_cols+\verb{_cleaned}.
}
\description{
Normalize addresses using data from \link{address_parse} and \link{address_geocode}.
}
