% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_exec_libpostal}
\alias{address_exec_libpostal}
\alias{address_parse_libpostal}
\alias{address_parse}
\title{address_exec_libpostal}
\usage{
address_exec_libpostal(addresses)

address_parse_libpostal(address_stream)

address_parse(address_stream)
}
\arguments{
\item{addresses}{character vector of addresses}

\item{address_stream}{data.table of addresses}
}
\value{
data.frame of parsed addresses, one row per vector i

data.table of addresses parsed, one per row in address_stream

data.table of addresses parsed
}
\description{
Parses addresses using libpostal and handles caching of already-parsed addresses so that they're only parsed once
}
\section{Functions}{
\itemize{
\item \code{address_exec_libpostal()}: execute address_parser and query it with a vector of addresses

\item \code{address_parse_libpostal()}: handle parsing by libpostal

}}
