% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_parse_libpostal}
\alias{address_parse_libpostal}
\title{address_parse_libpostal}
\usage{
address_parse_libpostal(address_stream)
}
\arguments{
\item{address_stream}{data.table of addresses}
}
\value{
data.table of addresses parsed
}
\description{
address_parse_libpostal
}
