% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_geocode.R
\name{address_reverse_census}
\alias{address_reverse_census}
\title{address_reverse_census}
\usage{
address_reverse_census(address_stream)
}
\arguments{
\item{address_stream}{data.table of addresses, must include \code{address_cols}}
}
\value{
data.table of addresses, one row per address in \code{address_stream}.  Contains only address_cols and
"state_fips", "county_fips", "census_tract", "census_block", "lat", "lon"
}
\description{
Gets census geography (tract/block/county/state) information for US addresses. Skips
lat/lon pairs that aren't in the US and which have already been forward geocoded and caches results.
}
