% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_stream}
\alias{address_stream}
\title{address_stream}
\usage{
address_stream(freshness = as.difftime(7, units = "days"))
}
\arguments{
\item{freshness}{data will be at least this fresh}
}
\value{
data.table of addresses and additional address-based features
}
\description{
Create a data.table of addresses, drawing from Tessitura TA_AUDIT_TRAIL and T_ADDRESS data, geocoding using the \link{tidygeocoder} package.
Tries each address up to six times, using \code{libpostal} parsing, \code{street1}, and \code{street2}, and the US census and openstreetmap geocoders.
Append census demographic and aggregate income data and iWave income data.
}
