% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collective_access.R
\name{ca_stream}
\alias{ca_stream}
\alias{ca_login}
\alias{ca_search}
\alias{ca_c}
\title{ca_stream}
\usage{
ca_stream(
  table_name = ca_table,
  ca_table,
  query = "*",
  features = NULL,
  base_url = config::get("tessistream.ca_base_url"),
  login = config::get("tessistream.ca_login")
)

ca_login(login, base_url)

ca_search(ca_table, query, base_url, bundles = NULL, login = NULL)

ca_c(...)
}
\arguments{
\item{table_name}{string}

\item{ca_table}{\href{1}{character} base CollectiveAccess table for the query (e.g. \code{ca_entities}, \code{ca_occurrences}, etc.)}

\item{query}{\href{1}{character} search query}

\item{features}{\link{list} of features to include in the dataset. List names will be used as the names of output columns, list
values identify the bundle (or a list of bundles that will concatenated) for the output column. Each bundle identifier can itself
be a list in order to specify additional parameters (e.g. \code{template}, \code{delimiter}, etc.)
\subsection{Example:}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
     "idno" = "idno",
     "name" = "preferred_labels",
     "season" = "ca_occurrences.hierarchy.preferred_labels",
     "date" = list("productionDate","screeningDate"),
     "artists" = list("ca_entities" = list(
       "template" = "<unit relativeTo='ca_entities' delimiter='|'>
                         ^preferred_labels (^relationship_typename)
                     </unit>")
     )
)
}\if{html}{\out{</div>}}
}}

\item{base_url}{\href{1}{character} CollectiveAccess base API endpoint
\href{https://manual.collectiveaccess.org/providence/developer/web_service_api.html#global-parameters}{ending in \code{service.php} or \code{service.php/json}}}

\item{login}{\href{1}{character} in the format of \code{username:password} for authenticating with CollectiveAccess}

\item{bundles}{\link{list} of bundles, as described in \href{https://manual.collectiveaccess.org/providence/developer/api_getting_data.html#bundle-specifiers}{the CollectiveAccess API documentation}}

\item{...}{lists/vectors to combine}
}
\value{
\link{character} authorization token

\link{data.table} with list columns corresponding to the items in \code{bundles}.
}
\description{
Create a dataset from a CollectiveAccess instance.
}
\section{Functions}{
\itemize{
\item \code{ca_login()}: login to a CollectiveAccess instance

\item \code{ca_search()}: execute a search query

\item \code{ca_c()}: Combine several input lists/vectors into a single flat vector, with no NULL,NA,or "" elements

}}
