% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_census.R
\name{census_data}
\alias{census_data}
\alias{census_get_data_all}
\title{census_data}
\usage{
census_data(census_variables, ...)

census_get_data_all(census_variables)
}
\arguments{
\item{census_variables}{data.table of years/datasets/variables to pull data for}

\item{...}{additional parameters to pass on to \link{address_cache}}
}
\value{
data.table of all census data and variable info
}
\description{
Loads and caches census data
}
\section{Functions}{
\itemize{
\item \code{census_data()}: Loads census data through cache

\item \code{census_get_data_all()}: Loads census data (does not cache)

}}
