% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_census.R
\name{census_race_features}
\alias{census_race_features}
\alias{census_sex_features}
\alias{census_age_features}
\alias{census_income_features}
\alias{census_features}
\title{census_features}
\usage{
census_race_features()

census_sex_features()

census_age_features()

census_income_features()

census_features()
}
\value{
data.table of features extracted from census datasets
}
\description{
Prepare features from raw census data
}
\section{Functions}{
\itemize{
\item \code{census_race_features()}: Prepares census race/ethnicity data

\item \code{census_sex_features()}: Prepares census sex data

\item \code{census_age_features()}: Prepares census age data

\item \code{census_income_features()}: Prepares census income data

}}
