% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_census.R
\name{census_get_data}
\alias{census_get_data}
\title{census_get_data}
\usage{
census_get_data(year, dataset, variables)
}
\arguments{
\item{year}{integer The year for which you are requesting data.}

\item{dataset}{character dataset to load from}

\item{variables}{character vector of variables to load}
}
\value{
data.table of census data
}
\description{
Wrapper around \link[tidycensus:get_decennial]{tidycensus::get_decennial} and \link[tidycensus:get_acs]{tidycensus::get_acs}
}
