% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_census.R
\name{census_race_features}
\alias{census_race_features}
\title{census_features}
\usage{
census_race_features()
}
\value{
data.table of features extracted from census datasets
}
\description{
Prepare features from raw census data
}
