% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_census.R
\name{census_variables}
\alias{census_variables}
\title{census_variables}
\usage{
census_variables()
}
\value{
data.table of census variables
}
\description{
Get a collection of basic census variables for sex, age, race, and income.
}
