% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collective_access.R
\name{collective_access_stream}
\alias{collective_access_stream}
\alias{collective_access_login}
\alias{collective_access_search}
\alias{collective_access_c}
\title{collective_access_stream}
\usage{
collective_access_stream(
  ca_table,
  table_name = ca_table,
  query = "*",
  features = NULL,
  base_url = config::get("tessistream")$collective_access_base_url,
  login = config::get("tessistream")$collective_access_login,
  batch_size = 100L,
  ...
)

collective_access_login(login, base_url)

collective_access_search(ca_table, query, base_url, login, bundles = NULL)

collective_access_c(...)
}
\arguments{
\item{ca_table}{\href{1}{character} base CollectiveAccess table for the query (e.g. \code{ca_entities}, \code{ca_occurrences}, etc.)}

\item{table_name}{\href{1}{character} output table name}

\item{query}{\href{1}{character} search query}

\item{features}{\link{list} of features to include in the dataset. List names will be used as the names of output columns, list
values identify the bundle (or a list of bundles that will concatenated) for the output column. Each bundle identifier can itself
be a list in order to specify additional parameters (e.g. \code{template}, \code{delimiter}, etc.)
\subsection{Example:}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
     "idno" = "idno",
     "name" = "preferred_labels",
     "season" = "ca_occurrences.hierarchy.preferred_labels",
     "date" = list("productionDate","screeningDate"),
     "artists" = list("ca_entities" = list(
       "template" = "<unit relativeTo='ca_entities' delimiter='|'>
                         ^preferred_labels (^relationship_typename)
                     </unit>")
     )
)
}\if{html}{\out{</div>}}
}}

\item{base_url}{\href{1}{character} CollectiveAccess base API endpoint
\href{https://manual.collectiveaccess.org/providence/developer/web_service_api.html#global-parameters}{ending in \code{service.php} or \code{service.php/json}}}

\item{login}{\href{1}{character} in the format of \code{username:password} for authenticating with CollectiveAccess}

\item{batch_size}{\href{1}{integer} number of records to request per query}

\item{...}{lists/vectors to combine}

\item{bundles}{\link{list} of bundles, as described in \href{https://manual.collectiveaccess.org/providence/developer/web_service_api.html#searching}{the CollectiveAccess API documentation}. The
name should be a bundle specifier and the value should be a list of parameters or an empty list.}
}
\value{
\link{character} authorization token

\link{data.table} with list columns corresponding to the items in \code{bundles}.
}
\description{
Create a dataset from a CollectiveAccess instance.
}
\section{Functions}{
\itemize{
\item \code{collective_access_login()}: login to a CollectiveAccess instance

\item \code{collective_access_search()}: execute a search query

\item \code{collective_access_c()}: Combine several input lists/vectors into a single flat character vector, with no NULL,NA,"null",or "" elements

}}
