% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_stream.R
\name{set_contribution_membership_match}
\alias{set_contribution_membership_match}
\alias{contribution_membership_match}
\title{contribution_membership_match}
\usage{
set_contribution_membership_match(contributions, memberships)

contribution_membership_match(contributions, memberships)
}
\arguments{
\item{contributions}{data.table of contributions, must include columns \verb{group_customer_no, campaign_category_desc, cont_amt, create_dt, cont_dt}}

\item{memberships}{data.table of memberships, must include columns \verb{group_customer_no, campaign_category_desc, memb_amt, AVC_amt, recog_amt, start_amt, create_dt, init_dt}}
}
\value{
matched data.table of contributions
}
\description{
Match memberships to contributions based on matching \code{create_dt}, \code{init_dt}/\code{cont_dt}, and alignment between
\verb{memb_amt + AVC_amt, recog_amt, start_amt} and the \code{cont_amt} of the contribution. Expected to fix 90\% of
unmatched contributions.
}
\section{Functions}{
\itemize{
\item \code{set_contribution_membership_match()}: updates \code{contributions$cust_memb_no}, intended to be called from within \link{contribution_membership_match}

}}
