% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{create_address_stream}
\alias{create_address_stream}
\title{create_address_stream}
\usage{
create_address_stream(freshness = as.difftime(7, units = "days"))
}
\arguments{
\item{freshness}{}
}
\description{
create_address_stream
}
