% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_stream.R
\name{email_stream}
\alias{email_stream}
\alias{email_data}
\title{email_stream}
\usage{
email_data(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[tessilake:read_sql]{tessilake::read_sql}}
  \describe{
    \item{\code{freshness}}{the returned data will be at least this fresh}
  }}
}
\description{
\itemize{
\item group_customer_no
\item timestamp : date of email event
\item event_type : Email
\item event_subtype : Open|Click|Unsubscribe|Hard Bounce|Forward
-* event_subtype2 : Onsale|Newsletter|Film|Fundraising-
\item campaign_no
\item appeal_no
\item source_no
\item extraction_desc
\item source_desc
\item address
\item domain
\item email[subtype]TimestampMin
\item email[subtype]TimestampMax
-* email[subtype][subtype2]Count-
}
}
\section{Functions}{
\itemize{
\item \code{email_data()}: load Tessitura email send/response data

}}
