% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_resilient.R
\name{make_resilient}
\alias{make_resilient}
\title{make_resilient}
\usage{
make_resilient(expr, num_tries = 10, sleep_secs = 10, default = NULL)
}
\arguments{
\item{expr}{expression to try to run}

\item{num_tries}{number of times to attempt to run \code{expr}}

\item{sleep_secs}{numbe of seconds to sleep between tries}

\item{default}{value to return if \code{expr} fails \code{num_tries} times}
}
\value{
the result of \code{expr} or \code{default} depending on whether \code{expr} ever ran successfully
}
\description{
Simple wrapper to retry an expression if it fails
}
