% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership_stream.R
\name{membership_stream}
\alias{membership_stream}
\title{membership_stream}
\usage{
membership_stream(control_period = years(4))
}
\arguments{
\item{control_period}{duration after the expiration of a membership to continue adding
control events for analysis}
}
\value{
\link[arrow:Table-class]{arrow::Table} of membership data
}
\description{
Generates a dataset of membership starts, ends and monthly 'controls' that
also keep track of changes to membership level.
Features:
\itemize{
\item \code{timestamp}
\item \code{customer_no}
\item \code{group_customer_no}
\item \code{event_type}: Membership
\item \code{event_subtype}: Start, End, and Control
\item \code{event_subtype2}: New, Renew, Reinstate
\itemize{
\item New: first membership ever
\item Renew: no lapse between memberships
\item Reinstate: membership starts after a lapse
}
\item \code{event_subtype3}: Current or Lapsed
\item \code{cust_memb_no}: the membership with the closest start/end
\item \code{cust_memb_no_prev}, \code{cust_memb_no_next}: the previous/next membership in the same organization or
outside of the organization if there are no more in the organization
\item \code{membership_count}: integer, number of memberships a customer has had
\item \code{membership_level}: character, the current membership level
\item \code{membership_amt}: double, the total value of memberships a customer has had
\item \code{membership_start_timestamp_min}, \code{membership_start_timestamp_max},
\code{membership_end_timestamp_min}, \code{membership_end_timestamp_max}: the first/last timestamp of a
membership start/end
}
}
