% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_orphan.R
\name{p2_add_tag}
\alias{p2_add_tag}
\title{p2_add_tag}
\usage{
p2_add_tag(contact, tag, dry_run = FALSE)
}
\arguments{
\item{contact}{integer P2 contact number}

\item{tag}{character, string for the tag}

\item{dry_run}{boolean, nothing will be changed in P2 if set to \code{TRUE}}
}
\description{
p2_add_tag
}
