% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_combine_jsons}
\alias{p2_combine_jsons}
\title{p2_combine_jsons}
\usage{
p2_combine_jsons(jsons)
}
\arguments{
\item{jsons}{list of list of \code{JSON} objects as data.tables, as returned by p2_json_to_datatable}
}
\value{
single \code{JSON} object as a list
}
\description{
Combine \code{JSON} objects returned by the Active Campaign API, concatenating each element with the same name.
}
