% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_db_update}
\alias{p2_db_update}
\title{p2_db_update}
\usage{
p2_db_update(data, table, overwrite = FALSE)
}
\arguments{
\item{data}{data.frame of data to write to the database}

\item{table}{table name}

\item{overwrite}{logical whether to delete and overwrite the existing table}
}
\value{
invisible
}
\description{
Write to the local P2 email database, either creating a new table or upserting into an existing one
}
