% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_email_map}
\alias{p2_email_map}
\title{p2_email_map}
\usage{
p2_email_map()
}
\value{
data.table of a mapping between email addresses and customer numbers
}
\description{
Constructs a many-to-many mapping between email addresses / P2 subscriber ids and Tessi customer numbers
}
