% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_orphan.R
\name{p2_execute_api}
\alias{p2_execute_api}
\title{p2_execute_api}
\usage{
p2_execute_api(
  url,
  object = list(),
  success_codes = c(200, 201, 202),
  method = "POST",
  api_key = keyring::key_get("P2_API"),
  dry_run = FALSE
)
}
\arguments{
\item{url}{character, endpoint url for the API}

\item{object}{list to be converted into JSON using \code{jsonlite::toJSON}}

\item{success_codes, }{integer vector of success codes returned by the API}

\item{method, }{character name of the \code{httr} function to call, usually \code{POST} or \code{PUT}}

\item{api_key}{Active Campaign API key, defaults to \code{keyring::key_get("P2_API")}}

\item{dry_run}{boolean, nothing will be changed in P2 if set to \code{TRUE}}
}
\value{
\code{TRUE} if success, \code{FALSE} if not
}
\description{
Sends a JSON \code{object} to a P2 API endpoint \code{url} using an \code{api_key} and returns T/F based on \code{success_codes}
}
