% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_get_activities}
\alias{p2_get_activities}
\title{p2_get_activities}
\usage{
p2_get_activities(contacts, start_dt = as.POSIXct("1970-01-01 00:00:00"))
}
\arguments{
\item{contacts}{vector of contact ids as integers or strings}

\item{start_dt}{POSIXct, only activities that have been created or updated since this date will be returned. (Note: side-loaded tables may have earlier entries.)}
}
\value{
\code{JSON} object of all email activities
}
\description{
p2_get_activities
}
