% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_get_contacts}
\alias{p2_get_contacts}
\title{p2_get_contacts}
\usage{
p2_get_contacts(start_dt = as.POSIXct("1970-01-01 00:00:00"))
}
\arguments{
\item{start_dt}{POSIXct, only contacts that have been created or updated since this date will be returned. (Note: side-loaded tables may have earlier entries.)}
}
\value{
single \code{JSON} object of all contacts
}
\description{
Load all contacts in P2 using Active Campaign API
}
