% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_json_to_datatable}
\alias{p2_json_to_datatable}
\title{p2_json_to_datatable}
\usage{
p2_json_to_datatable(json)
}
\arguments{
\item{json}{list of P2 JSON objects}
}
\description{
Convert P2 JSON objects to a data table
}
