% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_load}
\alias{p2_load}
\title{p2_load}
\usage{
p2_load(table, offset = 0, ...)
}
\arguments{
\item{table}{character, table to update}

\item{offset}{integer number of rows to offset from beginning of query}

\item{...}{additional parameters to pass on to modify_url}
}
\description{
Load p2 data from \code{api/3/{table}}, modified by arguments in \code{...} to the matching \code{table} in the local database
}
