% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_orphan.R
\name{p2_orphans}
\alias{p2_orphans}
\title{p2_orphans}
\usage{
p2_orphans()
}
\value{
data.table of orphaned accounts
}
\description{
Create a dataset of orphaned accounts in Prospect2 by comparing all accounts in Prospect2
against all accounts in Tessi. An orphan is an account that exists in P2, that originated
from Tessitura, but no longer matches an account with a primary email address in Tessitura.
}
