% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_orphan.R
\name{p2_orphans_report}
\alias{p2_orphans_report}
\title{p2_orphans_report}
\usage{
p2_orphans_report(...)
}
\arguments{
\item{...}{extra parameters passed on to tessi_changed_emails, default is \code{since = 0}}
}
\description{
sends an email containing a plot of recent orphans and a spreadsheet of all orphans
}
