% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_query_api}
\alias{p2_query_api}
\title{p2_query_api}
\usage{
p2_query_api(url, api_key = keyring::key_get("P2_API"), offset = 0)
}
\arguments{
\item{url}{Active Campaign API url to query}

\item{api_key}{Active Campaign API key, defaults to \code{keyring::key_get("P2_API")}}

\item{offset}{integer offset from the start of the query to return}
}
\value{
JSON object as a list
}
\description{
Parallel load from P2/Active Campaign API at \code{url} with key \code{api_key}. Loads pages of 100 records until it reaches the total.
}
