% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_orphan.R
\name{p2_resolve_orphan}
\alias{p2_resolve_orphan}
\title{p2_resolve_orphan}
\usage{
p2_resolve_orphan(from = NULL, to = NULL, customer_no = NULL, dry_run = FALSE)
}
\arguments{
\item{from}{character, email in P2 that needs to be changed}

\item{to}{character, email that will replace the \code{from} email}

\item{customer_no}{integer vector, customer numbers; one of which must match
the customer # field in P2 or the update will not be run}

\item{dry_run}{boolean, nothing will be changed in P2 if set to \code{TRUE}}
}
\value{
\code{TRUE} if update is run succesfully, \code{FALSE} if not.
}
\description{
Resolves orphan accounts in P2 based on the following tests:
\enumerate{
\item the email \code{from} exists in P2
\item the email \code{to} is not already used in P2
\item the customer # field in P2 matches a value passed in \code{customer_no}
}
}
\details{
It updates the \code{from} email in P2 to \code{to}, iff #1, #2 & #3 pass;
or it marks the account with a tag in P2 iff #1 and #3 pass, and #2 doesn't pass
}
