% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_stream}
\alias{p2_stream}
\title{p2_stream}
\usage{
p2_stream()
}
\value{
stream as a data.table
}
\description{
Combined dataset of email sends/clicks/opens/unsubscribes from Prospect2.
Features included are:
\itemize{
\item group_customer_no, customer_no
\item subscriberid : Prospect2 customer id
\item timestamp : date of email event
\item email : email address
\item event_type : "Email"
\item event_subtype : "Open", "Click", "Unsubscribe', "Hard Bounce", "Soft Bounce", etc.
\item campaignid, messageid, listid, linkid : Prospect2 internal ids
\item ip, ua, uasrc, referer, isread, times : Email open data
\item status : List subscription status
}
}
\section{Functions}{
\itemize{
\item \code{p2_stream()}: Update and build the p2 parquet files

}}
