% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_stream_enrich}
\alias{p2_stream_enrich}
\title{p2_stream_enrich}
\usage{
p2_stream_enrich(p2_stream)
}
\arguments{
\item{p2_stream}{stream data.table from p2_stream_build}
}
\value{
stream data.table with added descriptive columns
}
\description{
p2_stream_enrich
}
