% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_orphan.R
\name{p2_update_email}
\alias{p2_update_email}
\title{p2_update_email}
\usage{
p2_update_email(contact, email, dry_run = FALSE)
}
\arguments{
\item{contact}{integer P2 contact number}

\item{email}{character, new email for contact}

\item{dry_run}{boolean, nothing will be changed in P2 if set to \code{TRUE}}
}
\description{
Updates email for a contact on P2 (https://developers.activecampaign.com/reference/update-a-contact-new)
}
