% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessistream.R
\name{progress_expr}
\alias{progress_expr}
\title{progress_expr}
\usage{
progress_expr(expr, .progress)
}
\arguments{
\item{expr}{expression to evaluate}

\item{.progress}{progression function to call after each run of expr}
}
\value{
result of \code{expr}
}
\description{
Wrapper for incrementing progress after the end of \code{expr} evaluation.
}
