% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressr.R
\name{progressr_}
\alias{progressr_}
\title{progressr_}
\usage{
progressr_(n)
}
\arguments{
\item{n}{\link{integer} number of loops}
}
\value{
function to be called on each of n loops
}
\description{
Create a \link[progressr:progression]{progressr::progression} (or a dummy function)
}
