% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_utils.R
\name{setnafill}
\alias{setnafill}
\title{setnafill}
\usage{
setnafill(
  x,
  type = c("const", "locf", "nocb"),
  fill = NA,
  cols = seq_along(x),
  by = NA
)
}
\arguments{
\item{x}{data.table}

\item{type}{character, one of "\code{const}", "\code{locf}" or "\code{nocb}". Defaults to "\code{const}".}

\item{fill}{vector, value to be used to fill when \code{type=="const"}.}

\item{cols}{numeric or character vector specifying columns to be updated.}

\item{by}{character string of columns to group by}
}
\value{
filled data.table
}
\description{
Wrapper for data.table's setnafill for fast filling by group. Accepts non-numeric columns
}
\examples{
x <- data.table(a=NA,b=seq(1,10,10))
x[1,a:=1]
x[10,a:=3]
setnafill(x,"const",2,by="b")
setnafill(x,"locf",by="b")
setnafill(x,"nocb",by="b")

}
