% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_utils.R
\name{setunite}
\alias{setunite}
\title{setunite}
\usage{
setunite(data, col, ..., sep = "_", remove = TRUE, na.rm = FALSE)
}
\arguments{
\item{data}{data.table to act on}

\item{col}{The name of the new column, as a string or symbol.}

\item{...}{\verb{<tidy-select>} Columns to unite}

\item{sep}{Separator to use between values.}

\item{remove}{If \code{TRUE}, the default, remove input columns from output data frame.}

\item{na.rm}{If \code{TRUE}, missing values will be removed prior to uniting each value.}
}
\description{
Convenience function to paste together multiple columns into one. Thin wrapper around \link[tidyr:unite]{tidyr::unite}
}
