% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_upsert.R
\name{sqlite_upsert}
\alias{sqlite_upsert}
\title{sqlite_upsert}
\usage{
sqlite_upsert(con, table, data)
}
\arguments{
\item{con}{DBI connection object}

\item{table}{string name of table}

\item{data}{data.frame of data to update, rows will be matched on index columns}
}
\value{
invisibly
}
\description{
Updates a table in an sqlite database based on index columns
}
