% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream}
\alias{stream}
\alias{stream_chunk_write}
\alias{stream_window_features}
\title{stream}
\usage{
stream(
  streams = c("email_stream", "ticket_stream", "contribution_stream",
    "membership_stream", "ticketFuture_stream", "address_stream"),
  cols_match =
    "^(email|ticket|contribution|membership|ticket|address).+(amt|level|count|max|min)",
  windows = lapply(c(1, 7, 30, 90, 365), lubridate::days),
  rebuild = FALSE
)

stream_chunk_write(
  stream,
  stream_cols = setdiff(colnames(stream), c(by, "timestamp")),
  since = min(stream$timestamp) - 1,
  by = "group_customer_no",
  windows = NULL
)

stream_window_features(
  stream,
  stream_cols = setdiff(colnames(stream), c("group_customer_no", "timestamp")),
  windows = NULL,
  by = "group_customer_no"
)
}
\arguments{
\item{streams}{\link{character} vector of streams to combine}

\item{cols_match}{\link{character(1)} regular expression to use when matching columns to fill down}

\item{windows}{\link[lubridate:period]{lubridate::period} vector that determines the offsets used when constructing the windowed features.}

\item{rebuild}{\link{logical(1)} whether or not to rebuild the whole dataset (\code{TRUE}) or just append to the end of it (\code{FALSE})}
}
\value{
stream dataset as an \link[arrow:Table-class]{arrow::Table}
}
\description{
Combine all streams named in \code{streams} into a single dataset, filling down columns matched by \code{cols_match} and
optionally rebuilding the whole dataset if \code{rebuild} is \code{TRUE}. Windowed features are appended
}
\section{Functions}{
\itemize{
\item \code{stream_chunk_write()}: Fill down cols in \code{stream_cols} and add windowed features to \code{stream} for timestamps after \code{since}

\item \code{stream_window_features()}: construct windowed features for columns in \code{stream_cols}, using a list of \link[lubridate:period]{lubridate::period},
\code{windows} as offsets, and grouped by \code{by}

}}
