% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_utils.R
\name{stream_debounce}
\alias{stream_debounce}
\title{stream_debounce}
\usage{
stream_debounce(stream, ...)
}
\arguments{
\item{stream}{data.table in a stream format}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> columns to group by for debouncing}
}
\value{
de-bounced stream
}
\description{
Takes the last row for each group identified by columns identified in \code{...}
}
\examples{
stream <- data.table::data.table(
  x = 0:48, y = rep(0:4, 12)
)
data.table::setkey(stream, x)
stream_debounce(stream, "y")
}
