% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership_stream.R
\name{stream_effective_date}
\alias{stream_effective_date}
\title{stream_effective_date}
\usage{
stream_effective_date(stream, column, by = NULL)
}
\arguments{
\item{stream}{\code{data.table} with \code{timestamp}, and the columns named in \code{column}
and \code{by}, if defined.}

\item{column}{\code{character(1)} column name to make effective, must identify
a date or datetime column.}

\item{by}{\code{character(1)} optional column name to group rows by}
}
\value{
\code{data.table} with columns named by \code{column} and \code{by}
}
\description{
The effective date/time of a changing date/time column \code{dt} is the last
change before that change becomes effective, i.e. when \code{dt == timestamp}.
}
