% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_utils.R
\name{stream_from_audit}
\alias{stream_from_audit}
\title{stream_from_audit}
\usage{
stream_from_audit(table_name, ...)
}
\arguments{
\item{table_name}{character table name as in \code{tessilake::tessi_list_tables} \code{short_name} or \code{long_name}}

\item{...}{extra arguments passed on to \code{tessilake::read_tessi}}
}
\description{
Helper function to load data from the audit table
}
