% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/street_cleaner.R
\name{street_cleaner}
\alias{street_cleaner}
\title{street_cleaner}
\usage{
street_cleaner(list_no)
}
\arguments{
\item{list_no}{list number in Tessitura to pull data for}
}
\value{
data.table of cleaned addresses
}
\description{
Loads address info for a given list of customers and normalizes it with
\link{address_normalize}
}
