% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_stream.R
\name{survey_append_tessi}
\alias{survey_append_tessi}
\title{survey_append_tessi}
\usage{
survey_append_tessi(survey_data, tables, ...)
}
\arguments{
\item{tables}{character vector of tables to pass to \link[tessilake:read_tessi]{tessilake::read_tessi}}

\item{...}{expressions defining features that will be appended to survey data}

\item{survey_stream}{data.table of survey data}
}
\value{
data.table of survey data with features appended
}
\description{
survey_append_tessi
}
