% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_stream.R
\name{survey_cross}
\alias{survey_cross}
\title{survey_cross}
\usage{
survey_cross(survey_stream, ...)
}
\arguments{
\item{survey_stream}{data.table of survey data}

\item{...}{one or more regular expressions used to locate the questions to combine}
}
\value{
data.table of survey data, merged on \code{customer_hash} so that each question/answer is
a separate column, labelled \code{question.1}, \code{answer.1}, \code{question.2}, \code{answer.2}, etc. based on the
position of the selecting regular expression in \code{...}
}
\description{
survey_cross
}
