% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_stream.R
\name{survey_find_column}
\alias{survey_find_column}
\title{survey_find_column}
\usage{
survey_find_column(survey_data, .f, criterion = "max")
}
\arguments{
\item{survey_data}{data.frame of survey data}

\item{.f}{function to apply to each element of \code{survey_data}, should return a numeric score that is}

\item{criterion}{numeric or "max". If "max" then the first column with a maximum score is chosen, otherwise the first column with a value greater than
\code{criterion} is chosen.}
}
\value{
character column name
}
\description{
Find a column by applying \code{.f} to each element in the column, the results of which are summed, and then the column matching \code{criterion} is chosen.
}
