% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_stream.R
\name{survey_monkey}
\alias{survey_monkey}
\title{survey_monkey}
\usage{
survey_monkey(file)
}
\arguments{
\item{file}{filename of the file to process}
}
\value{
data.table stream of survey data. Columns are:
\itemize{
\item email
\item timestamp
\item question
\item subquestion
\item answer
}
}
\description{
Parser for Surveymonkey exports, handles:
}
\details{
\itemize{
\item identifying \code{email} and \code{timestamp} columns
\item appending survey name
\item labeling questions and subquestions
\item converting timestamps to POSIXct
}
}
