% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_stream.R
\name{survey_stream}
\alias{survey_stream}
\title{survey_stream}
\usage{
survey_stream()
}
\description{
\itemize{
\item anonymized using a hash so that joins can still be done post-hoc
\item timestamped so that trends can be visualized
}
\subsection{Features include:}{
\itemize{
\item customer_hash: string hash of customer_no
\item group_customer_hash: string hash of group_customer_no
\item timestamp: submission timestamp
\item question: text of the question
\item subquestion: text of the subquestion (or NA if none)
\item answer: text of the response
\item encoded_answer: embedding of the answer (e.g. integers for likert scale)
}
}
}
\details{
Simple dataset of all survey results.
}
\note{
There's no way to irreversibly anonymize this data and still allow post-hoc joins.
The secret in this case (the customer number) is stored openly in the database, the hashing
algorithm is explained here, and the number of possible customer numbers is small, so brute
forcing the mapping is trivial.

The goal is just to make it more difficult to extract customer information from this table
so that the end user \emph{knows} what they are doing.
}
