% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_orphan.R
\name{tessi_changed_emails}
\alias{tessi_changed_emails}
\title{tessi_changed_emails}
\usage{
tessi_changed_emails(since = Sys.Date() - 7)
}
\arguments{
\item{since}{date after which we look for changed emails}
}
\value{
data.table of changed emails with columns \code{old_value}, \code{new_value}, and \code{customer_no}
}
\description{
Returns list of primary emails changed since \code{since} date.
}
