% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessistream-package.R
\docType{package}
\name{tessistream-package}
\alias{tessistream}
\alias{tessistream-package}
\title{tessistream: Datasets for Analytics and Machine Learning from Tessitura}
\description{
| Builds timestamped datasets based on contributions, ticket sales, address changes, inventory for historical and predictive analysis.
}
\author{
\strong{Maintainer}: Sky Syzygy \email{ssyzygy@bam.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
