% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{write_xlsx}
\usage{
write_xlsx(data, xlsx_file)
}
\arguments{
\item{data}{data to write to the spreadsheet}

\item{xlsx_file}{filename to write to}
}
\value{
filename of written xlsx
}
\description{
Convenience wrapper around \code{openxlsx::write.xlsx} to handle some common formatting
\itemize{
\item row
}
}
