% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_cache}
\alias{address_cache}
\title{address_cache}
\usage{
address_cache(address_stream, cache_name, .function, ...)
}
\arguments{
\item{address_stream}{data.table of addresses}

\item{cache_name}{name of cache file passed to \code{tessilake::cache_write}}

\item{.function}{function to be called for processing}

\item{...}{additional options passed to \code{.function}}
}
\value{
data.table of addresses processed
}
\description{
Handles caching of already-processed addresses so that they're only processed once
}
