% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stream.R
\name{address_create_stream}
\alias{address_create_stream}
\title{address_create_stream}
\usage{
address_create_stream(freshness = as.difftime(7, units = "days"))
}
\arguments{
\item{freshness}{data will be at least this fresh}
}
\value{
data.table of addresses data at different points of time, no more than one
change per address per day
}
\description{
Creates address data with timestamps from TA_AUDIT_TABLE and T_ADDRESS data
}
