% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_stream.R
\name{contribution_membership_match}
\alias{contribution_membership_match}
\alias{set_contribution_membership_match}
\title{contribution_membership_match}
\usage{
contribution_membership_match(contributions, memberships)

set_contribution_membership_match(contributions, memberships)
}
\arguments{
\item{contributions}{data.table of contributions, must include columns \code{group_customer_no}, \code{ref_no}, \code{cust_memb_no}, \code{campaign_category_desc}, \code{cont_amt}, \code{create_dt}, \code{cont_dt}}

\item{memberships}{data.table of memberships, must include columns \code{group_customer_no}, \code{cust_memb_no}, \code{campaign_category_desc}, \code{memb_amt}, \code{AVC_amt}, \code{recog_amt}, \code{start_amt}, \code{create_dt}, \code{init_dt}}
}
\value{
data.table of contributions with \code{cust_memb_no} updated to match the appropriate membership in \code{memberships}
}
\description{
Match memberships to contributions based on matching \code{create_dt}, \code{init_dt}/\code{cont_dt}, and alignment between
\code{memb_amt + AVC_amt}, \code{recog_amt}, \code{start_amt} and the \code{cont_amt} of the contribution. Expected to fix 90\% of
unmatched contributions.
}
\section{Functions}{
\itemize{
\item \code{set_contribution_membership_match}: updates \code{contributions$cust_memb_no} for up to
one contribution per membership, intended to be called iteratively from within \link{contribution_membership_match}
}}

