% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_db_open}
\alias{p2_db_open}
\alias{p2_db_close}
\title{p2_db_open}
\usage{
p2_db_open(db_path = tessilake::cache_primary_path("p2.sqlite", "stream"))

p2_db_close()
}
\arguments{
\item{db_path}{path of the SQLite database}
}
\value{
invisible
}
\description{
Open the local P2 email database
}
\section{Functions}{
\itemize{
\item \code{p2_db_close}: Close the local P2 email database
}}

