% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_load}
\alias{p2_load}
\title{p2_load}
\usage{
p2_load(table, offset = 0, max_len = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{table}{character, table to update}

\item{offset}{integer offset from the start of the query to return}

\item{max_len}{integer maximum number of rows to load, defaults to
\link{p2_query_table_length}\verb{(url, api_key)} - \code{offset}.}

\item{overwrite}{logical, whether to overwrite the table in the database}

\item{...}{additional parameters to pass on to modify_url}
}
\description{
Load p2 data from \code{api/3/{table}}, modified by arguments in \code{...} to the matching \code{table} in the local database
}
