% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_orphan.R
\name{p2_orphans_report}
\alias{p2_orphans_report}
\title{p2_orphans_report}
\usage{
p2_orphans_report(freshness = 0)
}
\arguments{
\item{freshness}{difftime,    the returned data will be at least this fresh}
}
\description{
Sends an email containing a plot of recent orphans and a spreadsheet of all orphans.
}
