% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_stream.R
\name{p2_query_table_length}
\alias{p2_query_table_length}
\alias{p2_query_api}
\title{p2_query_api}
\usage{
p2_query_table_length(url, api_key = keyring::key_get("P2_API"))

p2_query_api(
  url,
  api_key = keyring::key_get("P2_API"),
  offset = 0,
  max_len = NULL
)
}
\arguments{
\item{url}{Active Campaign API url to query}

\item{api_key}{Active Campaign API key, defaults to \code{keyring::key_get("P2_API")}}

\item{offset}{integer offset from the start of the query to return}

\item{max_len}{integer maximum number of rows to load, defaults to
\link{p2_query_table_length}\verb{(url, api_key)} - \code{offset}.}
}
\value{
JSON object as a list
}
\description{
Parallel load from P2/Active Campaign API at \code{url} with key \code{api_key}. Loads pages of 100 records until it reaches the total.
}
\section{Functions}{
\itemize{
\item \code{p2_query_table_length}: get the length of the p2 API table at \code{url}
}}

