% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect2_execute_api.R
\name{p2_verb_thing}
\alias{p2_verb_thing}
\title{p2_verb_thing}
\usage{
p2_verb_thing(
  existing_things,
  new_things = NULL,
  verb,
  thing,
  dry_run = FALSE,
  .name = thing
)
}
\arguments{
\item{existing_things}{character vector of existing entity names}

\item{new_things}{list of objects to send to the API endpoint}

\item{verb}{HTTP verb to use for the API}

\item{thing}{character of entity to update (at "api/3/{thing}")}

\item{dry_run}{boolean, nothing will be changed in P2 if set to \code{TRUE}}

\item{.name}{character of name field that \code{existing_things} refers to}
}
\description{
p2_verb_thing
}
