% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_utils.R
\name{setnafill}
\alias{setnafill}
\alias{setnafill_const_simple}
\alias{setnafill_group}
\title{setnafill}
\usage{
setnafill(
  x,
  type = c("const", "locf", "nocb"),
  fill = NA,
  cols = seq_along(x),
  by = NA
)

setnafill_const_simple(x, type = "const", fill = NA, cols = seq_along(x))

setnafill_group(x, type = "locf", cols = seq_along(x), by = NA)
}
\arguments{
\item{x}{data.table}

\item{type}{character, one of "\code{const}", "\code{locf}" or "\code{nocb}". Defaults to "\code{const}".}

\item{fill}{vector, value to be used to fill when \code{type=="const"}.}

\item{cols}{numeric or character vector specifying columns to be updated.}

\item{by}{character string of columns to group by}
}
\value{
data.table, filled in-place
}
\description{
Wrapper for \link[data.table:nafill]{data.table::setnafill} for fast filling by group and non-numeric columns.
Displatches to setnafill_ functions based on column class and type of fill
}
\section{Functions}{
\itemize{
\item \code{setnafill_const_simple()}: Simple constant fill for factors and character

\item \code{setnafill_group()}: rolling join to quickly do setnafill by groups

}}
\examples{
x <- data.table::data.table(a = NA_character_, b = rep(seq(1, 2), each = 10))
x[1, a := "a"]
x[10, a := "x"]

y <- data.table::copy(x)
setnafill(y, type = "const", fill = "test", cols = "a")[]

y <- data.table::copy(x)
setnafill(y, type = "locf", by = "b")[]

y <- data.table::copy(x)
setnafill(y, type = "nocb", by = "b")[]

}
