% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream}
\alias{stream}
\alias{stream_chunk_write}
\alias{stream_window_features}
\title{stream}
\usage{
stream(
  streams = c("email_stream", "ticket_stream", "contribution_stream",
    "membership_stream", "ticket_future_stream", "address_stream"),
  fill_match =
    "^(email|ticket|contribution|membership|ticket|address).+(amt|level|count|max|min)",
  window_match = "^(email|ticket|contribution|membership|ticket|address).+(count|amt)$",
  rebuild = FALSE,
  ...
)

stream_chunk_write(
  stream,
  fill_cols = setdiff(colnames(stream), c(by, "timestamp")),
  window_cols = fill_cols,
  since = as.POSIXct(min(stream$timestamp) - 1),
  by = "group_customer_no",
  ...
)

stream_window_features(
  stream,
  window_cols = setdiff(colnames(stream), c("group_customer_no", "timestamp")),
  windows = NULL,
  by = "group_customer_no"
)
}
\arguments{
\item{streams}{\link{character} vector of streams to combine}

\item{fill_match}{\href{1}{character} regular expression to use when matching columns to fill down}

\item{window_match}{\href{1}{character} regular expression to use when matching columns to window}

\item{rebuild}{\href{1}{logical} whether or not to rebuild the whole dataset (\code{TRUE}) or just append to the end of it (\code{FALSE})}

\item{...}{not used}

\item{stream}{\link{data.table} data to process and write}

\item{fill_cols}{\link{character} columns to fill down}

\item{window_cols}{\link{character} columns to window}

\item{since}{\link{POSIXct} only the data with timestamps greater than \code{since} will be written}

\item{by}{\href{1}{character} column name to group by for filling down and windowing}

\item{windows}{\link[lubridate:period]{lubridate::period} vector that determines the offsets used when constructing the windowed features.}
}
\value{
stream dataset as an \link[arrow:Table-class]{arrow::Table}
}
\description{
Combine all streams named in \code{streams} into a single dataset, filling down columns matched by \code{cols_match} and
optionally rebuilding the whole dataset if \code{rebuild} is \code{TRUE}. Windowed features are appended using \code{windows}
as offsets.
}
\section{Functions}{
\itemize{
\item \code{stream_chunk_write()}: Fill down cols in \code{stream_cols} and add windowed features to \code{stream} for timestamps after \code{since}

\item \code{stream_window_features()}: construct windowed features for columns in \code{stream_cols}, using a list of \link[lubridate:period]{lubridate::period},
\code{windows} as offsets, and grouped by \code{by}

}}
