% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_stream.R
\name{stream_customer_history}
\alias{stream_customer_history}
\title{stream_customer_history}
\usage{
stream_customer_history(stream, before_date, pattern, ...)
}
\arguments{
\item{stream}{data.frameish stream}

\item{before_date}{POSIXct only look at customer history before this date}

\item{pattern}{character vector. If length > 1, the union of the matches is taken.}

\item{...}{
  Arguments passed on to \code{\link[tidyselect:starts_with]{tidyselect::matches}}
  \describe{
    \item{\code{ignore.case}}{If \code{TRUE}, the default, ignores case when matching
names.}
    \item{\code{perl}}{Should Perl-compatible regexps be used?}
  }}
}
\description{
Loads the last row from \code{stream} per \code{group_customer_no} before \code{before_date} and returns columns
matching \code{pattern}
}
