% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_utils.R
\name{stream_customer_history}
\alias{stream_customer_history}
\title{stream_customer_history}
\usage{
stream_customer_history(
  stream,
  by,
  before = as.POSIXct("2100-01-01"),
  pattern = ".",
  ...
)
}
\arguments{
\item{stream}{data.frameish stream}

\item{by}{character column name to group by}

\item{before}{POSIXct only look at customer history before this date}

\item{pattern}{character vector. If length > 1, the union of the matches is taken.}

\item{...}{
  Arguments passed on to \code{\link[tidyselect:starts_with]{tidyselect::matches}}
  \describe{
    \item{\code{ignore.case}}{If \code{TRUE}, the default, ignores case when matching
names.}
    \item{\code{perl}}{Should Perl-compatible regexps be used?}
  }}
}
\description{
Loads the last row from \code{stream} after sorting by \code{timestamp} and grouping by the entries in columns \code{by},
but only looking at timestamps before \code{before} and returning only columns matching \code{pattern}
}
