% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_clean.R
\name{street_cleaner}
\alias{street_cleaner}
\title{street_cleaner}
\usage{
street_cleaner(list_no)
}
\arguments{
\item{list_no}{list number in Tessitura to pull data for}
}
\value{
data.table of cleaned addresses
}
\description{
Convenience function for loading address info for a given list of customers and normalizing it by
running it through \link{address_normalize}
}
